pyproject.toml
[project]
name = "lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "pyyaml>=6.0.2",
]
